requirements.txt
Pillow>=10.0.0
opencv-python
scikit-image
numpy
matplotlib
scipy
streamlit
streamlit-option-menu
numpy>=1.24
opencv-python>=4.9.0